src/backend/pyproject.toml
#
# Meet package
#
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "meet"
version = "0.1.7"
authors = [{ "name" = "DINUM", "email" = "dev@mail.numerique.gouv.fr" }]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Framework :: Django",
    "Framework :: Django :: 5",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]
description = "A simple video and phone conferencing tool, powered by LiveKit"
keywords = ["Django", "Contacts", "Templates", "RBAC"]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3==1.35.34",
    "Brotli==1.1.0",
    "celery[redis]==5.4.0",
    "django-configurations==2.5.1",
    "django-cors-headers==4.4.0",
    "django-countries==7.6.1",
    "django-parler==2.3",
    "redis==5.1.1",
    "django-redis==5.4.0",
    "django-storages[s3]==1.14.4",
    "django-timezone-field>=5.1",
    "django==5.1.1",
    "djangorestframework==3.15.2",
    "drf_spectacular==0.27.2",
    "dockerflow==2024.4.2",
    "easy_thumbnails==2.10",
    "factory_boy==3.3.1",
    "gunicorn==23.0.0",
    "jsonschema==4.23.0",
    "june-analytics-python==2.3.0",
    "markdown==3.7",
    "nested-multipart-parser==1.5.0",
    "psycopg[binary]==3.2.3",
    "PyJWT==2.9.0",
    "python-frontmatter==1.1.0",
    "requests==2.32.3",
    "sentry-sdk==2.15.0",
    "url-normalize==1.4.3",
    "WeasyPrint>=60.2",
    "whitenoise==6.7.0",
    "mozilla-django-oidc==4.0.1",
    "livekit-api==0.7.0",
]

[project.urls]
"Bug Tracker" = "https://github.com/numerique-gouv/meet/issues/new"
"Changelog" = "https://github.com/numerique-gouv/meet/blob/main/CHANGELOG.md"
"Homepage" = "https://github.com/numerique-gouv/meet"
"Repository" = "https://github.com/numerique-gouv/meet"

[project.optional-dependencies]
dev = [
    "django-extensions==3.2.3",
    "drf-spectacular-sidecar==2024.7.1",
    "freezegun==1.5.1",
    "ipdb==0.13.13",
    "ipython==8.27.0",
    "pyfakefs==5.6.0",
    "pylint-django==2.5.5",
    "pylint==3.2.7",
    "pytest-cov==5.0.0",
    "pytest-django==4.9.0",
    "pytest==8.3.3",
    "pytest-icdiff==0.9",
    "pytest-xdist==3.6.1",
    "responses==0.25.3",
    "ruff==0.6.5",
    "types-requests==2.32.0.20240914",
]

[tool.setuptools]
packages = { find = { where = ["."], exclude = ["tests"] } }
zip-safe = true

[tool.distutils.bdist_wheel]
universal = true

[tool.ruff]
exclude = [
    ".git",
    ".venv",
    "build",
    "venv",
    "__pycache__",
    "*/migrations/*",
]
line-length = 88


[tool.ruff.lint]
select = [
    "B", # flake8-bugbear
    "BLE", # flake8-blind-except
    "C4", # flake8-comprehensions
    "DJ", # flake8-django
    "I", # isort
    "PLC", # pylint-convention
    "PLE", # pylint-error
    "PLR", # pylint-refactoring
    "PLW", # pylint-warning
    "RUF100", # Ruff unused-noqa
    "RUF200", # Ruff check pyproject.toml
    "S", # flake8-bandit
    "SLF", # flake8-self
    "T20", # flake8-print
]
ignore= ["DJ001", "PLR2004"]

[tool.ruff.lint.isort]
section-order = ["future","standard-library","django","third-party","meet","first-party","local-folder"]
sections = { meet=["core"], django=["django"] }

[tool.ruff.lint.per-file-ignores]
"**/tests/*" = ["S", "SLF"]

[tool.pytest.ini_options]
addopts = [
    "-v",
    "--cov-report",
    "term-missing",
    # Allow test files to have the same name in different directories.
    "--import-mode=importlib",
]
python_files = [
    "test_*.py",
    "tests.py",
]
